requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=1.4.0=pyhd8ed1ab_0
appdirs=1.4.4=pyh9f0ad1d_0
autograd=1.5=pyhd8ed1ab_0
autoray=0.3.1=pyhd8ed1ab_0
blas=1.0=mkl
brotli=1.1.0=hd590300_0
brotli-bin=1.1.0=hd590300_0
brotli-python=1.1.0=py311hb755f60_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=hd590300_0
ca-certificates=2023.7.22=hbcca054_0
cachetools=5.3.1=pyhd8ed1ab_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py311h409f033_3
charset-normalizer=3.2.0=pyhd8ed1ab_0
chex=0.1.5=py311h06a4308_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.1.1=py311h9547e67_0
cvxopt=1.3.2=py311h2b3fd1d_0
cvxpy=1.3.2=py311h38be061_0
cvxpy-base=1.3.2=py311h320fe9a_0
cycler=0.11.0=pyhd8ed1ab_0
cyrus-sasl=2.1.28=h52b45da_1
dataclasses=0.8=pyhc8e2a94_3
dbus=1.13.18=hb2f20db_0
dill=0.3.7=pyhd8ed1ab_0
dm-haiku=0.0.10=pyhd8ed1ab_0
dm-tree=0.1.8=py311h46cbc50_1
dsdp=5.8=hd9d9efa_1203
ecos=2.0.11=py311hcb2cf0a_0
expat=2.5.0=hcb278e6_1
fftw=3.3.10=nompi_hc118613_108
filelock=3.12.4=pyhd8ed1ab_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.42.1=py311h459d7ec_0
freetype=2.12.1=h4a9f257_0
future=0.18.3=pyhd8ed1ab_0
giflib=5.2.1=h0b41bf4_3
glib=2.69.1=he621ea3_2
glpk=5.0=h445213a_0
gmp=6.2.1=h58526e2_0
gsl=2.7=he838d99_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.9.0=py311hdd6beaf_0
hdf5=1.12.1=h2b7332f_3
icu=58.2=hf484d3e_1000
idna=3.4=pyhd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46305
jax=0.3.25=py311h06a4308_0
jaxlib=0.3.25=py311h6a678d5_2
jinja2=3.1.2=pyhd8ed1ab_1
jmp=0.0.4=pyhd8ed1ab_0
joblib=1.3.2=pyhd8ed1ab_0
jpeg=9e=h0b41bf4_3
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=py311h9547e67_0
krb5=1.20.1=h81ceb04_0
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libabseil=20230802.1=cxx17_h59595ed_0
libblas=3.9.0=1_h86c2bf4_netlib
libbrotlicommon=1.1.0=hd590300_0
libbrotlidec=1.1.0=hd590300_0
libbrotlienc=1.1.0=hd590300_0
libcblas=3.9.0=5_h92ddd45_netlib
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcups=2.4.2=h2d74bed_1
libcurl=8.2.1=h251f7ec_0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h516909a_1
libevent=2.1.12=hf998b51_1
libexpat=2.5.0=hcb278e6_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_1
libgfortran-ng=13.2.0=h69a702a_1
libgfortran5=13.2.0=ha4646dd_1
liblapack=3.9.0=5_h92ddd45_netlib
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.52.0=h2d74bed_1
libosqp=0.6.3=h59595ed_0
libpng=1.6.39=h5eee18b_0
libpq=12.15=hdbd6064_1
libprotobuf=3.20.3=he621ea3_0
libqdldl=0.1.5=h27087fc_1
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=13.2.0=h7e041cc_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h1daa5a0_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.16=hd590300_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hcbfbd50_0
libxslt=1.1.37=h873f0b0_0
llvm-openmp=12.0.1=h4bd325d_1
lz4-c=1.9.4=hcb278e6_0
markupsafe=2.1.3=py311h459d7ec_0
matplotlib=3.8.0=py311h38be061_0
matplotlib-base=3.8.0=py311h54ef318_0
metis=5.1.0=h59595ed_1007
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
mpfr=4.2.0=hb012696_0
mpmath=1.3.0=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
networkx=3.1=pyhd8ed1ab_0
ninja=1.11.1=h924138e_0
numpy=1.23.5=py311h7d28db0_0
openssl=3.1.3=hd590300_0
opt_einsum=3.3.0=pyhd8ed1ab_1
optax=0.1.7=pyhd8ed1ab_0
osqp=0.6.3=py311h320fe9a_1
packaging=23.1=pyhd8ed1ab_0
pandas=2.1.0=py311h320fe9a_0
patsy=0.5.3=pyhd8ed1ab_0
pcre=8.45=h9c3ff4c_0
pennylane=0.32.0=py311h38be061_2
pennylane-core=0.32.0=py311h38be061_2
pennylane-lightning-core=0.32.0=py311hd08793c_2
pillow=9.4.0=py311h6a678d5_0
pip=23.2.1=py311h06a4308_0
ply=3.11=py_1
pthread-stubs=0.4=h36c2ea0_1001
pycparser=2.21=pyhd8ed1ab_0
pyparsing=3.1.1=pyhd8ed1ab_0
pyqt=5.15.7=py311h6a678d5_0
pyqt5-sip=12.11.0=py311h6a678d5_0
pysocks=1.7.1=pyha2e5f31_6
python=3.11.5=h955ad1f_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.11=2_cp311
pytorch=2.0.1=cpu_py311h6d93b4c_0
pytz=2023.3.post1=pyhd8ed1ab_0
pyyaml=6.0=py311h5eee18b_1
qdldl-python=0.1.5.post2=py311h8b32b4d_0
qt-main=5.15.2=h7358343_9
qt-webengine=5.15.9=hbbf29b9_6
qtwebkit=5.212=h3fafdc1_5
readline=8.2=h5eee18b_0
requests=2.31.0=pyhd8ed1ab_0
rustworkx=0.13.1=py311h1069841_0
scikit-learn=1.3.0=py311hc009520_0
scipy=1.11.2=py311h64a7726_1
scs=3.2.3=py311h18430ed_0
seaborn=0.13.0=hd8ed1ab_0
seaborn-base=0.13.0=pyhd8ed1ab_0
semantic_version=2.10.0=pyhd8ed1ab_0
setuptools=68.0.0=py311h06a4308_0
sip=6.6.2=py311h6a678d5_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h5eee18b_0
statsmodels=0.14.0=py311h1f0f07a_2
suitesparse=5.10.1=h9e50725_1
sympy=1.12=pyh04b8f61_3
tabulate=0.9.0=pyhd8ed1ab_1
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.2.0=pyha21a80b_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
toolz=0.12.0=pyhd8ed1ab_0
tornado=6.3.3=py311h459d7ec_0
tqdm=4.66.1=pyhd8ed1ab_0
typing-extensions=4.8.0=hd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023c=h04d1e81_0
urllib3=2.0.4=pyhd8ed1ab_0
wheel=0.38.4=py311h06a4308_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7f98852_2
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
